+++ Runtime/Internal/IStoreInternalEventUseCase.cs
@@ -1,0 +1,7 @@
+namespace AffiseAttributionLib.Internal
+{
+    internal interface IStoreInternalEventUseCase
+    {
+        void StoreInternalEvent(InternalEvent internalEvent);
+    }
+}