+++ Runtime/Internal/InternalEventsParams.cs
@@ -1,0 +1,9 @@
+namespace AffiseAttributionLib.Internal
+{
+    internal static class InternalEventsParams
+    {
+        internal const string INTERNAL_EVENTS_DIR_NAME = "affise-internal-events";
+        internal const int INTERNAL_EVENTS_SEND_COUNT = 100;
+        internal const int INTERNAL_EVENTS_STORE_TIME = 7 * 24 * 60 * 60 * 1000;
+    }
+}