+++ Runtime/Internal/InternalParameters.cs
@@ -1,0 +1,10 @@
+namespace AffiseAttributionLib.Internal
+{
+    internal static class InternalParameters
+    {
+        internal const string AFFISE_INTERNAL_EVENT_ID = "internal_event_id";
+        internal const string AFFISE_INTERNAL_EVENT_NAME = "internal_event_name";
+        internal const string AFFISE_INTERNAL_EVENT_TIMESTAMP = "internal_event_timestamp";
+        internal const string AFFISE_INTERNAL_EVENT_DATA = "internal_event_data";
+    }
+}