+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
 
+## [1.7.10] - 2026-07-14
+
+### Added
+
+- Internal event `SessionStart`
+
+### Changed
+
+- Update native iOS to [`1.7.10`](https://github.com/affise/affise-mmp-sdk-ios/blob/1.7.10/CHANGELOG.md)
+- Update native Android to [`1.7.10`](https://github.com/affise/affise-mmp-sdk-android/blob/v1.7.10/CHANGELOG.md)
+
 ## [1.7.9] - 2026-06-22
 
 ### Added
 
 - Provider `ADVERTISER_TRACKING_ENABLED`
 - Provider `APPLICATION_TRACKING_ENABLED`
-- Provider `CARRIER`
 - Provider `CPU_CORES`
 - Provider `SCREEN_WIDTH`
 - Provider `SCREEN_HEIGHT`
@@ -596,6 +606,7 @@
 - Update native iOS to `1.6.20`
 - Update native Android to `1.6.22`
 
+[1.7.10]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.9...1.7.10
 [1.7.9]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.8...1.7.9
 [1.7.8]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.7...1.7.8
 [1.7.7]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.6...1.7.7